Super Cache: Guarantee a unique value for the page secret
Previously the admin would have to set a unique value in `wp-cache-config.php`, which was easy to overlook. If they didn't do that, then the `donotcachepage` parameter could be used in a DDoS attack.

This makes it much more obvious, and also falls back to disabling the functionality for backwards-compatibility.

Fixes https://hackerone.com/reports/185914
Props https://hackerone.com/e3amn2l