Regolith Mail: Detect commands with `command` for... some reason.
Supposedly this is the best practice for determining if a command exists. There's an explanation of why at https://unix.stackexchange.com/a/85250/18886, but to be honest, I don't have the patience to wade through that to extract the relevant point. Instead, I'm just gonna trust the Stack Exchange hive mind :)