Regolith Mail: Allow plugins to override `From` name and `Reply-To`.
When Core or plugins send mail, they know more about the context than `Regolith\Mail` can, so they should be able to set the `From` name and `Reply-To` headers. The `From` address should always be controlled by `Regolith\Mail`, though, to ensure that it isn't forged, which would result in messages being flagged as spam.